+++ addons/zylann.hterrain/src/resources.gen.h
@@ -1,4 +1,0 @@
-
-// This is a generated file. Do not edit.
-
-extern const char *s_default_shader_code;